remove unused template-style analyzers
